Multi threading experiment
Spawn a new thread for each directory encountered.

Because inodes is a constantly updating hashmap it was removed for this
example.

Performance was bad. Despite the code does less work as the inode check
had to be removed. It was roughly half the speed of the single threaded
version. I infer that the cost of creating a new thread for each dir
far outweighs the benefit of using multiple threads.

It is possible that creating a small thread pool and letting
a limited number of threads run would perform better. However we would
still have to solve the problem of the inode hashmap which must be
updatable by all threads.